backend/requirements.txt

aioice==0.9.0
aiortc==1.5.0
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
av==10.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dnspython==2.4.2
fastapi==0.101.1
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
google-crc32c==1.5.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
ifaddr==0.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pyee==11.0.0
pylibsrtp==0.8.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.159
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
